fully test cover semaphore and mock process.kill
as it turns out the reason kill couldn't be redefined was because it had
a different scope. kill is actually a singleton method which needs to be
overridden in a bit of a different way.

allow mocking kill as either process-running or process-dead. also add
a new test for the now possible process-running scenario we couldn't
test before